requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.10.0
beautifulsoup4==4.13.5
Brotli==1.1.0
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.1.8
cssselect==1.3.0
decorator==4.4.2
fastapi==0.116.1
feedfinder2==0.0.4
feedparser==6.0.11
ffmpy==0.6.1
filelock==3.19.1
fsspec==2025.9.0
google-api-core==2.25.1
google-api-python-client==2.181.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
gradio==5.44.1
gradio_client==1.12.1
groovy==0.1.2
gspread==6.2.1
gTTS==2.5.4
h11==0.16.0
hf-xet==1.1.9
httpcore==1.0.9
httplib2==0.30.0
httpx==0.28.1
huggingface-hub==0.34.4
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
jieba3k==0.35.1
Jinja2==3.1.6
joblib==1.5.2
lxml==6.0.1
lxml_html_clean==0.4.2
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
moviepy==1.0.3
newspaper3k==0.2.8
nltk==3.9.1
numpy==2.3.2
oauth2client==4.1.3
oauthlib==3.3.1
orjson==3.11.3
packaging==25.0
pandas==2.3.2
pillow==11.3.0
proglog==0.1.12
proto-plus==1.26.1
protobuf==6.32.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
regex==2025.9.1
requests==2.32.5
requests-file==2.1.0
requests-oauthlib==2.0.0
rich==14.1.0
rsa==4.9.1
ruff==0.12.12
safehttpx==0.1.6
semantic-version==2.10.0
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
starlette==0.47.3
tinysegmenter==0.3
tldextract==5.3.0
tomlkit==0.13.3
tqdm==4.67.1
typer==0.17.4
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.35.0
websockets==15.0.1
